requirements.txt
astroid==3.3.5
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
dill==0.3.9
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
idna==3.10
isort==5.13.2
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mccabe==0.7.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
packaging==24.1
pillow==11.0.0
platformdirs==4.3.6
pylint==3.3.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.3.0
six==1.16.0
sympy==1.13.1
tomlkit==0.13.2
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
urllib3==2.2.3
